refactor: InquiryRequestDto 및 InquiryResponseDto 패키지 이동
