+++ src/main/java/com/YOGIITSU/dto/RequestDto/InquiryRequestDto.java
@@ -1,4 +1,4 @@
-package com.YOGIITSU.dto;
+package com.YOGIITSU.dto.RequestDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
